(feat): Affichage point sur tableau

+++ includes/cube3d.h
@@ -6,7 +6,7 @@
 /*   By: gpollast <gpollast@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/10 11:07:42 by gpollast          #+#    #+#             */
-/*   Updated: 2025/11/11 17:41:26 by gpollast         ###   ########.fr       */
+/*   Updated: 2025/11/12 17:59:38 by gpollast         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,8 +25,8 @@ typedef	struct	s_map
 
 typedef struct	s_player
 {
-	double	x;
-	double	y;
+	int	x;
+	int	y;
 }	t_player;
 
 typedef struct s_game
